PR comments, moving more stuff into SDK.
