+++ packages/server/src/sdk/users/index.ts
@@ -1,7 +1,5 @@
-import * as sync from "./sync"
 import * as utils from "./utils"
 
 export default {
-  ...sync,
   ...utils,
 }
